Dummy data for broken pollen json
The json string from Claritin is no longer available. Replaced call to
Claritin with data used for testing. Pollen info is now static but it
works for demonstration purposes. All non-pollen is still live.

It's been 3 months since we last worked on this and I'm ahead/behind by
2 and 3 commits respectively somehow so who knows what's going to happen
when I sync this. Wish me luck!